split click commands into multiple files
