--- gcal_nest/__main__.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-'''
-This module wraps command_line.main().
-'''
-
-# Imports #####################################################################
-from .cli import main
-
-
-if __name__ == '__main__':
-    main()
